requirements.txt
psycopg2-binary
tabulate
colorama